feat: remplacer l'icône du site par un SVG et extraire les fonctions utilitaires pour la gestion des arbres de fichiers
